+++ cmake/compiler_utils.cmake
@@ -33,7 +33,23 @@ function(get_version version_str)
     set(PATCH_VERSION ${TMP_PATCH_VERSION} PARENT_SCOPE)
 endfunction()
 
+function(is_dir_empty dir)
+    file(GLOB RESULT dir)
+    list(LENGTH RESULT num_files)
+    if(num_files EQUAL 0)
+        set(dir_is_empty TRUE)
+    else()
+        set(dir_is_empty FALSE)
+    endif()
+endfunction()
+
+
 function(get_muparserx_source_code)
+    is_dir_empty(${PROJECT_SOURCE_DIR}/src/third-party/headers/muparserx)
+    if(NOT dir_is_empty)
+        message(STATUS "MuparserX library source code already exists")
+        return()
+    endif()
     find_package(Git QUIET)
     if(GIT_FOUND AND EXISTS "${PROJECT_SOURCE_DIR}/.git")
         # if we have cloned the sources, muparserx is a submodule, so we need
@@ -91,10 +107,14 @@ function(uncompress_muparsersx_lib)
         if(CXX11_ABI EQUAL "0")
             set(MUPARSER_ABI_PREFIX oldabi_)
         endif()
+        if(CMAKE_SYSTEM_PROCESSOR MATCHES "ppc64le")
+            set(MUPARSER_ARC_POSTFIX ".ppc64le")
+        endif()
+        
         set(PLATFORM "linux")
     endif()
 
-    execute_process(COMMAND ${CMAKE_COMMAND} -E tar "xvfj" "${AER_SIMULATOR_CPP_SRC_DIR}/third-party/${PLATFORM}/lib/${MUPARSER_ABI_PREFIX}muparserx.7z"
+    execute_process(COMMAND ${CMAKE_COMMAND} -E tar "xvfj" "${AER_SIMULATOR_CPP_SRC_DIR}/third-party/${PLATFORM}/lib/${MUPARSER_ABI_PREFIX}muparserx${MUPARSER_ARC_POSTFIX}.7z"
             WORKING_DIRECTORY  "${AER_SIMULATOR_CPP_SRC_DIR}/third-party/${PLATFORM}/lib/")
     set(MUPARSERX_LIB_PATH "${AER_SIMULATOR_CPP_SRC_DIR}/third-party/${PLATFORM}/lib" PARENT_SCOPE)
 endfunction()