+++ qiskit/providers/aer/utils/noise_transformation.py
@@ -31,12 +31,12 @@ import itertools
 import logging
 import numpy
 
-from qiskit.providers.aer.noise.errors import QuantumError
-from qiskit.providers.aer.noise import NoiseModel
-from qiskit.providers.aer.noise.noiseerror import NoiseError
-from qiskit.providers.aer.noise.errors.errorutils import single_qubit_clifford_instructions
 from qiskit.quantum_info.operators.channel import Kraus
 from qiskit.quantum_info.operators.channel import SuperOp
+from ..noise.errors import QuantumError
+from ..noise.noise_model import NoiseModel
+from ..noise.noiseerror import NoiseError
+from ..noise.errors.errorutils import single_qubit_clifford_instructions
 
 logger = logging.getLogger(__name__)